# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar App OS API
  description: >
    REST API for the Stellar App OS reforestation platform. Provides endpoints
    for tree planting verification, carbon credit management, escrow operations,
    NFT minting, compliance reporting, and more.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Farm Credit / Stellar App OS
    url: https://github.com/Farm-credit/stellar-app-os
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://stellar-app-os.vercel.app
    description: Production
security: []
...
paths:
...
  /api/planting/photo:
    post:
      operationId: postPlantingPhoto
      tags: [Planting]
      summary: Upload and verify planting photo
      description: >
        Uploads a photo to IPFS, verifies GPS proximity to assigned tree,
        and returns the IPFS CID.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo, lat, lon, farmerId]
              properties:
                photo:
                  type: string
                  format: binary
                lat:
                  type: string
                lon:
                  type: string
                farmerId:
                  type: string
                treeId:
                  type: string
                sponsorEmail:
                  type: string
                sponsorName:
                  type: string
                region:
                  type: string
                  default: unknown
      responses:
        "201":
          description: Photo uploaded and verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  ipfsCid:
                    type: string
                  ipfsUrl:
                    type: string
                  gatewayUrl:
                    type: string
                  s3Key:
                    type: string
                  encryptedGps:
                    type: object
                  hash:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: GPS distance verification failed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - type: object
                    properties:
                      distanceMeters:
                        type: number
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/admin/credits/airdrop:
    get:
      operationId: getAdminAirdropPreview
      tags: [Admin]
      summary: Preview airdrop recipients (admin only)
      security:
        - adminAuth: []
      parameters:
        - name: platformLaunchDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: creditsPerSponsor
          in: query
          required: true
          schema:
            type: number
            minimum: 0
            exclusiveMinimum: true
      responses:
        "200":
          description: Airdrop preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AirdropPreview"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: postAdminAirdrop
      tags: [Admin]
      summary: Execute airdrop (admin only)
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [creditsPerSponsor, projectId, platformLaunchDate]
              properties:
                creditsPerSponsor:
                  type: number
                  minimum: 0
                  exclusiveMinimum: true
                projectId:
                  type: string
                platformLaunchDate:
                  type: string
                  format: date
      responses:
        "200":
          description: Airdrop executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AirdropResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ──────────────────────────────────────────────
  # NFT
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /api/auth/login
    adminAuth:
      type: apiKey
      in: header
      name: x-admin-secret
      description: Admin secret key
    webhookCronAuth:
      type: http
      scheme: bearer
      description: WEBHOOK_CRON_SECRET token
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    AirdropPreview:
      type: object
      properties:
        recipients:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              walletAddress:
                type: string
              email:
                type: string
              joinedAt:
                type: string
                format: date-time
        totalCredits:
          type: integer
        cutoffDate:
          type: string
          format: date
    AirdropResult:
      type: object
      properties:
        totalQueued:
          type: integer
        recipients:
          type: array
          items:
            type: object
            properties:
              walletAddress:
                type: string
              status:
                type: string
                example: queued
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...